feat: merge APIs into single FastAPI app
