--- services/insight/README.md
@@ -1,84 +1,0 @@
-# Insight Researcher Service
-
-This service calls the OpenAI API to generate short research reports and summary insights.
-It runs at `http://localhost:8083` when using Docker Compose.
-
-## Endpoints
-
-- `GET /health` – liveness probe returning `{ "status": "ok" }`.
-- `GET /ready` – always returns `{ "ready": true }`.
-- `POST /generate-insights` – body `{ "text": "notes" }` returns `{ "markdown": "..." }`.
-- `POST /insight` – body `{ "url": "https://example.com", "industry": "SaaS", "pain_point": "Slow onboarding", "stack": [{"category": "analytics", "vendor": "GA4"}] }`
-  returns `{ "markdown": "..." }`.
-- `POST /research` – body `{ "topic": "AI" }` returns `{ "markdown": "..." }`.
-- `POST /postprocess-report` – body `{ "report": {...} }` returns the same report
-  plus base64-encoded downloads.
-- `POST /insight-and-personas` – body `{ "url": "https://example.com", "industry": "SaaS", "pain_point": "Slow onboarding", "stack": [{"category": "analytics", "vendor": "GA4"}], "evidence_standards": "Use peer-reviewed data", "credibility_scoring": "1-5", "deliverable_guidelines": "Plain language", "audience": "CTO", "preferences": "Focus on OSS" }`
-  returns { "insight": {"actions": [...], "evidence": "..."}, "personas": [{"id": "P1"}], "cms_manual": "WordPress", "degraded": false }. Insight and persona prompts run concurrently. The `evidence` field summarizing findings is always included. If the persona response is empty, placeholder company and technology personas are created with all attributes set to "unknown". The gateway will return a timeout after 20s if the insight service is slow. The optional fields fine‑tune how the report is generated.
-- `GET /metrics` – usage counters for requests and data gaps.
-
-## Environment variables
-
-- `OPENAI_API_KEY` – required key for the OpenAI client.
-- `OPENAI_MODEL` – chat model name used by `call_openai_with_retry`
-  (default `gpt-4`).
-- `OPENAI_MAX_TOKENS` – maximum tokens returned by the LLM
-  (default `800`).
-- `MACRO_SECTION_CAP` – maximum number of macro sections returned by `/research`.
-
-### Normalized insight schema
-
-The insight service replies with an object containing an `insight` field. This
-field follows a consistent shape regardless of prompt details. An `evidence`
-summary is always included, and placeholder personas with `"unknown"` fields
-are returned when persona data is missing:
-
-```json
-{
-  "evidence": "short summary text",
-  "actions": [
-    {"id": "1", "title": "Action", "reasoning": "why", "benefit": "result"}
-  ],
-  "personas": [
-    {"id": "P1", "name": "Buyer", "role": "unknown", "goal": "unknown", "challenge": "unknown"}
-  ],
-  "degraded": false
-}
-```
-
-Additional keys may be included but these core fields are always present. The
-frontend's `InsightCard` component expects this structure.
-
-## Example usage
-
-```bash
-curl -X POST http://localhost:8083/research \
-  -H 'Content-Type: application/json' \
-  -d '{"topic": "latest AI trends"}'
-```
-
-Expected response snippet:
-
-```json
-{
-  "markdown": "..."
-}
-```
-
-Example (insight and personas):
-
-```bash
-curl -X POST http://localhost:8083/insight-and-personas \
-  -H 'Content-Type: application/json' \
-  -d '{"url": "https://example.com", "industry": "SaaS", "pain_point": "Slow onboarding", "stack": [{"category": "analytics", "vendor": "GA4"}], "evidence_standards": "Use peer-reviewed data", "credibility_scoring": "1-5", "deliverable_guidelines": "Plain language", "audience": "CTO", "preferences": "Focus on OSS"}'
-```
-
-Expected response snippet:
-
-```json
-{
-  "insight": { "actions": [], "evidence": "..." },
-  "personas": [{"id": "P1", "name": "Buyer", "role": "unknown", "goal": "unknown", "challenge": "unknown"}],
-  "degraded": false
-}
-```
