+++ tests/test_app_service.py
@@ -1,30 +1,16 @@
 from fastapi.testclient import TestClient
 
-from services.gateway.app import app as gateway_app
-from services.martech.app import app as martech_app
+from main import app as main_app
 from services.property.app import app as property_app
-from services.insight.app import app as insight_app
 
 
-def test_load_gateway():
-    client = TestClient(gateway_app)
-    r = client.get("/health")
-    assert r.status_code == 200
-
-
-def test_load_martech():
-    client = TestClient(martech_app)
+def test_load_main():
+    client = TestClient(main_app)
     r = client.get("/health")
     assert r.status_code == 200
 
 
 def test_load_property():
     client = TestClient(property_app)
-    r = client.get("/health")
-    assert r.status_code == 200
-
-
-def test_load_insight():
-    client = TestClient(insight_app)
     r = client.get("/health")
     assert r.status_code == 200
